refactor: 重构 transition 组件
